+++ CHANGELOG.md
@@ -14,4 +14,4 @@
 
 ### Feat
 
-- 优化 text 模型构建 prompt 的逻辑
+- 优化 text 模型构建 prompt 的逻辑，去除点击登录无用操作，下线 text-curie-001, text-babbage-001, text-ada-001 较为古老的模型，增加 text-davinci-002 模型，优化 REDME.md 文档样式